Cache me not

+++ bin/ccache.sh
@@ -1,14 +1,16 @@
 #!/bin/sh
-if [ -d /Sites/ephect ];
+
+pwd
+
+if [ ! -d ./cache ] || [ ! -d ./runtime ];
 then
-    cd /Sites/ephect;
-fi
-if [ -d /Users/david/Sites/CodePhoenixOrg/SDK/php/ephect ];
-then
-    cd /Users/david/Sites/CodePhoenixOrg/SDK/php/ephect;
+    echo "Be sure you are at the root of your Ephect site."
+    exit 1;
 fi
 
 sudo rm -rf cache
 sudo rm -rf runtime
+rm php_errors.log
+rm src/public/php_errors.log
 echo Cache cleared
 echo
